LPS-93519 Move com.liferay.portal.editor.configuration tests to the module as Arquillian tests
